+++ Cargo.toml
@@ -29,6 +29,7 @@ insta = { version = "1.47.2", features = ["yaml"] }
 etcetera = "0.11.0"
 jarl-core = { path = "crates/jarl-core" }
 jarl-lsp = { path = "crates/jarl-lsp" }
+jarl-semantic = { path = "crates/jarl-semantic" }
 path-absolutize = "3.1.1"
 rayon = "1.12.0"
 regex = { version = "1.11.1", default-features = false, features = ["std"] }
