+++ crates/jarl-core/Cargo.toml
@@ -29,6 +29,9 @@ oak_index.workspace = true
 oak_index_vec.workspace = true
 url.workspace = true
 
+# Per-file semantic facts (used by DFG-based lints)
+jarl-semantic.workspace = true
+
 # Serialization and data structures
 serde.workspace = true
 toml.workspace = true
